Ajuste na validação

+++ Core/Application/Importacao/Commands/Inputs/ProdutoCommand.cs
@@ -7,7 +7,7 @@ namespace Core.Application.Importacao.Commands.Inputs
         public int Id { get; set; }
         public int Linha{ get; set; }
         public string Nome { get; set; }
-        public long Quantidade { get; set; }
+        public int Quantidade { get; set; }
         public DateTime DataEntrega { get; set; }
         public decimal Valor { get; set; }
     }
