+++ Core/Application/Importacao/Commands/Models/ImportacaoProdutoEntity.cs
@@ -2,8 +2,8 @@
 {
     public class ImportacaoProdutoEntity
     {
-        public long ImportacaoId { get; set; }
-        public long ProdutoId { get; set; }
+        public int ImportacaoId { get; set; }
+        public int ProdutoId { get; set; }
 
         public ImportacaoEntity Importacao { get; set; }
         public ProdutoEntity Produto { get; set; }
